requirements.txt
torch==2.1.1 
torchvision==0.16.1  
albumentations==1.3.1 
Pillow==9.3.0 
numpy==1.24.1
rasterio==1.3.9 
pyyaml==6.0.1
matplotlib==3.8.2
